refactor(service-worker): review and enhance
