--- packages/service-worker/src/lib/logger.ts
@@ -1,3 +1,0 @@
-import { createLogger } from '@alwatr/logger';
-
-export const logger = /* @__PURE__ */ createLogger('service-worker');
