+++ tsconfig.json
@@ -2,7 +2,9 @@
   "extends": "@nexim/typescript-config/tsconfig.json",
   "compilerOptions": {
     "noEmit": true,
-    "types": ["node"]
+    "types": [
+      "node"
+    ]
   },
   "files": [],
   "exclude": [],
@@ -15,6 +17,12 @@
     },
     {
       "path": "./packages/alpine"
+    },
+    {
+      "path": "./packages/service-worker"
+    },
+    {
+      "path": "./packages/snackbar"
     }
   ]
 }
